+++ CHANGELOG.md
@@ -32,6 +32,8 @@ Breaking changes should be prefixed by `[**BREAKING**]` (without the quotes), to
 
 ### Added
 
+- Generators:
+  - `rails generate federails:copy_factories` copies FactoryBot factories in 'spec/factories', with `federails_` prefix
 - Actors:
   - added `sync!` method to update distant actors: `Actor.find(1).sync!`
   - added `on_federails_undelete_requested` callback